+++ cmd/config/config.go
@@ -1,0 +1,16 @@
+// Copyright Tharsis Labs Ltd.(Evmos)
+// SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
+
+package config
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/evmos/os/types"
+)
+
+// SetBip44CoinType sets the global coin type to be used in hierarchical deterministic wallets.
+func SetBip44CoinType(config *sdk.Config) {
+	config.SetCoinType(types.Bip44CoinType)
+	config.SetPurpose(sdk.Purpose)                  // Shared
+	config.SetFullFundraiserPath(types.BIP44HDPath) //nolint: staticcheck
+}