+++ server/config/config.go
@@ -16,8 +16,6 @@ import (
 	errorsmod "cosmossdk.io/errors"
 	"github.com/cosmos/cosmos-sdk/server/config"
 	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/crypto-org-chain/cronos/memiavl"
-	memiavlcfg "github.com/crypto-org-chain/cronos/store/config"
 )
 
 const (
@@ -98,27 +96,6 @@ const (
 
 	// DefaultGasAdjustment value to use as default in gas-adjustment flag
 	DefaultGasAdjustment = 1.2
-
-	// ============================
-	//           MemIAVL
-	// ============================
-
-	// DefaultMemIAVLEnable is the default value that defines if memIAVL is enabled
-	DefaultMemIAVLEnable = false
-
-	// DefaultZeroCopy is the default value that defines if
-	// the zero-copied slices must be retained beyond current block's execution
-	// the sdk address cache will be disabled if zero-copy is enabled
-	DefaultZeroCopy = false
-
-	// DefaultAsyncCommitBuffer value to use as default for the size of
-	// asynchronous commit queue when using memIAVL
-	DefaultAsyncCommitBuffer = 0
-
-	// DefaultSnapshotKeepRecent default value for how many old snapshots
-	// (excluding the latest one) should be kept after new snapshots
-	// when using memIAVL
-	DefaultSnapshotKeepRecent = 1
 )
 
 var evmTracers = []string{"json", "markdown", "struct", "access_list"}
@@ -131,8 +108,6 @@ type Config struct {
 	EVM     EVMConfig     `mapstructure:"evm"`
 	JSONRPC JSONRPCConfig `mapstructure:"json-rpc"`
 	TLS     TLSConfig     `mapstructure:"tls"`
-
-	MemIAVL MemIAVLConfig `mapstructure:"memiavl"`
 }
 
 // EVMConfig defines the application configuration values for the EVM.
@@ -194,59 +169,6 @@ type TLSConfig struct {
 	CertificatePath string `mapstructure:"certificate-path"`
 	// KeyPath the file path for the key .pem file
 	KeyPath string `mapstructure:"key-path"`
-}
-
-// MemIAVLConfig defines the configuration for memIAVL.
-type MemIAVLConfig struct {
-	memiavlcfg.MemIAVLConfig
-}
-
-// AppConfig helps to override default appConfig template and configs.
-// return "", nil if no custom configuration is required for the application.
-func AppConfig(denom string) (string, interface{}) {
-	// Optionally allow the chain developer to overwrite the SDK's default
-	// server config.
-	customAppConfig := DefaultConfig()
-
-	// The SDK's default minimum gas price is set to "" (empty value) inside
-	// app.toml. If left empty by validators, the node will halt on startup.
-	// However, the chain developer can set a default app.toml value for their
-	// validators here.
-	//
-	// In summary:
-	// - if you leave srvCfg.MinGasPrices = "", all validators MUST tweak their
-	//   own app.toml config,
-	// - if you set srvCfg.MinGasPrices non-empty, validators CAN tweak their
-	//   own app.toml to override, or use this default value.
-	//
-	// In evmos, we set the min gas prices to 0.
-	if denom != "" {
-		customAppConfig.Config.MinGasPrices = "0" + denom
-	}
-
-	customAppTemplate := config.DefaultConfigTemplate +
-		DefaultEVMConfigTemplate +
-		memiavlcfg.DefaultConfigTemplate
-
-	return customAppTemplate, *customAppConfig
-}
-
-// DefaultConfig returns server's default configuration.
-func DefaultConfig() *Config {
-	defaultSDKConfig := config.DefaultConfig()
-	defaultSDKConfig.API.Enable = DefaultAPIEnable
-	defaultSDKConfig.GRPC.Enable = DefaultGRPCEnable
-	defaultSDKConfig.GRPCWeb.Enable = DefaultGRPCWebEnable
-	defaultSDKConfig.Rosetta.Enable = DefaultRosettaEnable
-	defaultSDKConfig.Telemetry.Enabled = DefaultTelemetryEnable
-
-	return &Config{
-		Config:  *defaultSDKConfig,
-		EVM:     *DefaultEVMConfig(),
-		JSONRPC: *DefaultJSONRPCConfig(),
-		TLS:     *DefaultTLSConfig(),
-		MemIAVL: *DefaultMemIAVLConfig(),
-	}
 }
 
 // DefaultEVMConfig returns the default EVM configuration
@@ -377,30 +299,21 @@ func (c TLSConfig) Validate() error {
 	return nil
 }
 
-// DefaultMemIAVLConfig returns the default MemIAVL configuration
-func DefaultMemIAVLConfig() *MemIAVLConfig {
-	return &MemIAVLConfig{memiavlcfg.MemIAVLConfig{
-		Enable:             DefaultMemIAVLEnable,
-		ZeroCopy:           DefaultZeroCopy,
-		AsyncCommitBuffer:  DefaultAsyncCommitBuffer,
-		SnapshotKeepRecent: DefaultSnapshotKeepRecent,
-		SnapshotInterval:   memiavl.DefaultSnapshotInterval,
-		CacheSize:          memiavlcfg.DefaultCacheSize,
-	}}
-}
+// DefaultConfig returns server's default configuration.
+func DefaultConfig() *Config {
+	defaultSDKConfig := config.DefaultConfig()
+	defaultSDKConfig.API.Enable = DefaultAPIEnable
+	defaultSDKConfig.GRPC.Enable = DefaultGRPCEnable
+	defaultSDKConfig.GRPCWeb.Enable = DefaultGRPCWebEnable
+	defaultSDKConfig.Rosetta.Enable = DefaultRosettaEnable
+	defaultSDKConfig.Telemetry.Enabled = DefaultTelemetryEnable
 
-// Validate returns an error if the MemIAVL configuration fields are invalid.
-func (c MemIAVLConfig) Validate() error {
-	// AsyncCommitBuffer can be -1, which means synchronous commit
-	if c.AsyncCommitBuffer < -1 {
-		return errors.New("AsyncCommitBuffer cannot be negative")
+	return &Config{
+		Config:  *defaultSDKConfig,
+		EVM:     *DefaultEVMConfig(),
+		JSONRPC: *DefaultJSONRPCConfig(),
+		TLS:     *DefaultTLSConfig(),
 	}
-
-	if c.CacheSize < 0 {
-		return errors.New("CacheSize cannot be negative")
-	}
-
-	return nil
 }
 
 // GetConfig returns a fully parsed Config object.
@@ -424,10 +337,6 @@ func (c Config) ValidateBasic() error {
 
 	if err := c.TLS.Validate(); err != nil {
 		return errorsmod.Wrapf(errortypes.ErrAppConfig, "invalid tls config value: %s", err.Error())
-	}
-
-	if err := c.MemIAVL.Validate(); err != nil {
-		return errorsmod.Wrapf(errortypes.ErrAppConfig, "invalid memIAVL config value: %s", err.Error())
 	}
 
 	return c.Config.ValidateBasic()