+++ types/hdpath.go
@@ -1,5 +1,6 @@
 // Copyright Tharsis Labs Ltd.(Evmos)
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
+
 package types
 
 import (
@@ -18,7 +19,7 @@ type (
 	HDPathIterator func() ethaccounts.DerivationPath
 )
 
-// HDPathIterator receives a base path as a string and a boolean for the desired iterator type and
+// NewHDPathIterator receives a base path as a string and a boolean for the desired iterator type and
 // returns a function that iterates over the base HD path, returning the string.
 func NewHDPathIterator(basePath string, ledgerIter bool) (HDPathIterator, error) {
 	hdPath, err := ethaccounts.ParseDerivationPath(basePath)