+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This changelog was created using the `clu` binary
 
 ### Improvements
 
+- (cli) [#23](https://github.com/evmos/os/pull/23) Add client, cmd and version packages.
 - (server) [#22](https://github.com/evmos/os/pull/22) Add server implementation.
 - (rpc) [#21](https://github.com/evmos/os/pull/21) Add RPC and indexer types.
 - (types) [#20](https://github.com/evmos/os/pull/20) Add crypto and encoding packages.